make fast_ig_tools more reusable (cherry picked from commit a78462e)
