+++ src/fast_ig_tools/ig_trie_compressor.cpp
@@ -14,6 +14,8 @@ namespace po = boost::program_options;
 
 #include "fast_ig_tools.hpp"
 #include "ig_trie_compressor.hpp"
+#include "utils.hpp"
+
 using fast_ig_tools::Trie;
 
 #include <seqan/seq_io.h>
